# public/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Good Game Pickems API",
    "version": "1.0.0",
    "description": "API for accessing Good Game Pickems predictions, matches, and leaderboards",
    "contact": {
      "name": "API Support",
      "url": "https://goodgamepickems.com/developer"
    }
  },
  "servers": [
    {
      "url": "https://goodgamepickems.com/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/matches": {
      "get": {
        "summary": "List matches",
        "description": "Get a list of matches with optional filters",
        "parameters": [
          {
            "name": "game_type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["csgo", "lol", "valorant"]
            },
            "description": "Filter by game type"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["upcoming", "live", "completed"]
            },
            "description": "Filter by match status"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/predictions": {
      "get": {
        "summary": "List predictions",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "match_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create prediction",
        "security": [
          {
            "ApiKeyAuth": ["write"]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrediction"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Prediction created"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/leaderboard": {
      "get": {
        "summary": "Get leaderboard",
        "parameters": [
          {
            "name": "game_type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["csgo", "lol", "valorant"]
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["today", "week", "month", "all_time"],
              "default": "all_time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            }
          }
        }
      }
    },
    "/overlay": {
      "get": {
        "summary": "Get overlay data for streaming",
        "description": "Returns user stats and recent predictions formatted for stream overlays",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 5,
              "minimum": 1,
              "maximum": 20
            },
            "description": "Number of recent predictions to include"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverlayResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "Match": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "home_team": {
            "type": "string"
          },
          "away_team": {
            "type": "string"
          },
          "match_date": {
            "type": "string",
            "format": "date-time"
          },
          "game_type": {
            "type": "string",
            "enum": ["csgo", "lol", "valorant"]
          },
          "home_score": {
            "type": "integer",
            "nullable": true
          },
          "away_score": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "Prediction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "match_id": {
            "type": "string",
            "format": "uuid"
          },
          "picked_winner": {
            "type": "string"
          },
          "home_score": {
            "type": "integer",
            "nullable": true
          },
          "away_score": {
            "type": "integer",
            "nullable": true
          },
          "points": {
            "type": "integer"
          }
        }
      },
      "LeaderboardEntry": {
        "type": "object",
        "properties": {
          "rank": {
            "type": "integer"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "total_points": {
            "type": "integer"
          },
          "accuracy": {
            "type": "string"
          }
        }
      },
      "MatchesResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Match"
            }
          },
          "meta": {
            "type": "object"
          }
        }
      },
      "PredictionsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Prediction"
            }
          }
        }
      },
      "LeaderboardResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntry"
            }
          }
        }
      },
      "CreatePrediction": {
        "type": "object",
        "required": ["match_id", "picked_winner"],
        "properties": {
          "match_id": {
            "type": "string",
            "format": "uuid"
          },
          "picked_winner": {
            "type": "string"
          },
          "home_score": {
            "type": "integer"
          },
          "away_score": {
            "type": "integer"
          }
        }
      },
      "OverlayResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "total_points": {
                    "type": "integer"
                  },
                  "correct_picks": {
                    "type": "integer"
                  },
                  "total_picks": {
                    "type": "integer"
                  }
                }
              },
              "predictions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Prediction"
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing API key",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "UNAUTHORIZED"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "FORBIDDEN"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}